Support rejoining to the game
Client side is a bit buggy (old player messages are shown in the blue boxes, and played cards are a total disaster).

On server side game history is now saved.
Players request the history when they connect to the game. When they are offline, events are not sent to the client (but Player object still accepts the events)
Login names are currently ignored, players have preconfigured names.