Add cool animation for card playing
The played card shrinks from the player's hand.
And the taking animation is played reversed when playing a card. I tried to make this animation much earlier, but it didn't work. Now I figured out what the problem was: the when the animation was started, the card view's size was zero (because the card image was null), and it broke the animation. Now the card view has constant size.