Use only one thread on server for games
Now there are one reader and writer threads per client, one tcp listener thread, and one game thread. This change elimnated the possible race condition at message handling (they are on the game thread now), and the unnecessary threads per game. I'm starting to be more and more statisfied with the quality of the code. There are lot more to do, but I'm certainly getting good quality upgrades :D

Actually, when I finished this, I didn't have time to test it, because I went to play tarokk (in real life, strange enough, isn't it?).
And when I came back, I totally forgot I made this change. And everything worked. When I realized, that I made this change, and it didn't cause any problems, I was really suprised and happy :D.