Add test for bidding
Osvalt László made an excel table with the all possible biddings.

There are several failures in the test, mostly caused by missing passz in the test cases, or different invitation rules, but it's ok.
One minor bug in the program: players are expected to passz even if it is not necessary (after 3, 0; or 3, 2, 1). But it's acceptable.