Fix randomly deleting game
This bug was with us a long long time, the game was deleted during play, but it happened rarely, and I always just ignored it, and wanted to fix it later. Until this bug became very frequent in a game. After a bit thinking and trying, I found out the cause of the bug.
Actually I always knew that it should be caused by some UI shit, but I couldn't really believe this. But I was right. The hidden game list fragment got the unhandled click events from the gameplay fragment, and this caused the games to be deleted when the user clicked on that area of the game fragment. I'm really happy now, this was the worst bug in this project so far, and now it is fixed :DD.

The fix is: the fragments should be replaced by the other, and not just being covered. Now everything seems to be perfect :).