Do not force played card view height
Let it be as high as it wants (wrap content)
The reason why this ugly hack was used (forcing card height to be 1.6 times the width), was animations needed the height. But now with RELATIVE_TO_SELF type animations this problem goes away.

Wow, it's good. The code is a bit cleaner now.