Add kibic chat
Chat is not an action anymore, it's a different message. The chat event still exists but its parameter is a user id, not a seat.
Chats are moved to a different database table. This complicates game session loading a bit, but allows us to save the chat even in lobbies, and is also more logical, because chat does not modify the game state as actions do.