Game: remove points argument from constructor
Now Game does not track points, since we might have more than 4 players each having their own points.
When game is ended, GameSession gets the points from Game, and adds it to the total points of the players, which is stored in the Player class.
The code became much nicer this way. We removed one more argument from Game constructor :).