Make chat work with old (3.0) app also
Due to coronavirus google play review times increased a lot, so users cannot update the app to the latest version. The chat protocol is changed, so chat is not working with the old app anymore.

This commit makes old chat protocol to work. Kibic chat is still invisible, but at least the player chat works as usual.