Fix PagatSasUltimo and Zaroparos relation
The current implementation of PagatSasUltimo was wrong, and the rules in my head was also wrong :D.

The new rule is: if zaroparos is announced, there is no silent pagat sas ultimo.

(The previous implementation disabled silent ultimos only when zaroparos was successful, and also when zaroparos was just "silently" successful, which was clearly a bug).