Fix hkp announcing for caller player
Changing the rule that hkp is required for all cards (not just XX) introduced a bug, that the caller player has to announce hkp if he folded the card he called. This led to infinite amount of hkp from bots, overloading the server.

Also checking whether hkp is announced is unncecessary, as it's not even stored, checking the contra level of the game is the thing we need.