ListDeferredResultService: return result only after commit
I felt that ListDeferredResultService needed one more step to make it
very robust, I think this was that. Now the result is only returned
after the transaction is completed that generated the result.

The problem was at the end of the game: after the last action was returned,
the game state query was too fast, and didn't return the updated state
(it sould have returned 410, but the new game was not created yet).
Now the result is delayed, so the state query returns the updated state.

We also get some robustness by not returning the result if the transatction
was rolled back.

Finally, a really nice impelentation of this service :).